+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.3.0
+
+Limit the size of logged `variables` for `meta.graphql` in the middleware logger
+via `maxGraphQLVariablesLength`. It's possible that `variables` is a big object that could bloat the payload.
+
 ## 1.2.0
 
 Add handling for `requestId` ([#18](https://github.com/molindo/molindo-node-logger/pull/18))